requirements.txt
streamlit
pandas 
bs4
requests
times
numpy
tfcausalimpact
matplotlib
openpyxl